+++ web-frontend/modules/database/plugin.js
@@ -62,6 +62,13 @@ import {
   RowUpdatedWebhookEventType,
   RowDeletedWebhookEventType,
 } from '@baserow/modules/database/webhookEventTypes'
+import {
+  ImageFilePreview,
+  AudioFilePreview,
+  VideoFilePreview,
+  PDFBrowserFilePreview,
+  GoogleDocFilePreview,
+} from '@baserow/modules/database/filePreviewTypes'
 import { APITokenSettingsType } from '@baserow/modules/database/settingsTypes'
 
 import tableStore from '@baserow/modules/database/store/table'
@@ -339,6 +346,13 @@ export default (context) => {
     'formula_type',
     new BaserowFormulaSingleSelectType(context)
   )
+
+  // File preview types
+  app.$registry.register('preview', new ImageFilePreview(context))
+  app.$registry.register('preview', new AudioFilePreview(context))
+  app.$registry.register('preview', new VideoFilePreview(context))
+  app.$registry.register('preview', new PDFBrowserFilePreview(context))
+  app.$registry.register('preview', new GoogleDocFilePreview(context))
 
   registerRealtimeEvents(app.$realtime)
 }